POP OAUTH2 for Microsoft
This commit is based on two patches provided by M. Stoeckl in April 2022.
See URL's below.

A note of caution: this fixes the POP issues for Microsoft mail servers,
and it is now tested by at least three persons in three different mail
environments, but this code might break POP with OAUTH2 for Gmail and/or
Yahoo if either requires the original behavior.

Microsoft requires OAUTH2 authentication to be split over two lines, so
instead of

 POP> AUTH XOAUTH2 B64

they require

 POP> AUTH XOAUTH2
 POP< +
 POP> B64

This includes the changes in the patches provided in tickets 4587 and 4590

https://www.thewildbeast.co.uk/claws-mail/bugzilla/show_bug.cgi?id=4587
https://www.thewildbeast.co.uk/claws-mail/bugzilla/show_bug.cgi?id=4590